Ajuste de tipo de coluna de preço de Room

+++ DTOs/HotelSearchDto.cs
@@ -5,7 +5,7 @@ public class HotelSearchDto
     public DateTime? MininumDate { get; set; }
     public DateTime? MaximumDate { get; set; }
     public string? CityName { get; set; }
-    public double? MinimumPrice { get; set; }
-    public double? MaximumPrice { get; set; }
+    public decimal? MinimumPrice { get; set; }
+    public decimal? MaximumPrice { get; set; }
     public short? MinimumRate { get; set; }
 }
